Rename NVIDIA controller module
Rename gpu_controller.rs to nvidia_controller.rs so the NVIDIA and AMD backend modules use matching names.
